backend/requirements.txt
# Django Core
Django==4.2.9
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==25.1

# Authentication & Authorization
djangorestframework-simplejwt==5.3.0
djoser==2.2.0

# Database
psycopg2-binary==2.9.9
django-extensions==3.2.3

# File Storage & Cloud
django-storages==1.14.2
boto3==1.34.34
Pillow==10.2.0

# Background Tasks
django-q==1.3.9
redis==3.5.3

# API Documentation
drf-spectacular==0.27.1

# Development Tools
django-debug-toolbar==4.2.0
factory-boy==3.3.0

# Security & Validation
cryptography==43.0.3
pycryptodome==3.19.0

# Hardware/OBD Interface
pyserial==3.5
python-can==4.3.1

# Utilities
python-decouple==3.8
requests==2.31.0
celery==5.3.4

# Testing
pytest-django==4.8.0
pytest-cov==4.1.0

# Production
gunicorn==21.2.0
whitenoise==6.6.0

# Payment Processing
stripe==7.12.0
paypal-checkout-serversdk==1.0.1

# Revenue & Analytics
django-money==3.4.1
py-moneyed==3.0

# Enhanced Validation
jsonschema==4.21.1
pydantic==2.5.3 